🔀 merge로 인한 mindmap 초기화 코드 리팩토링
소켓설정해서 하던 것을 useProject로 이동
