+++ client/src/hooks/useProject/index.ts
@@ -1,7 +1,9 @@
 import useRoomId from 'hooks/useRoomId';
 import { useEffect } from 'react';
 import { useRecoilState, useSetRecoilState } from 'recoil';
+import { mindmapState } from 'recoil/mindmap';
 import { labelListState, projectIdState, sprintListState, userListState } from 'recoil/project';
+import { IMindNodes } from 'types/mindmap';
 import { project } from 'utils/api';
 
 const useProject = () => {
@@ -9,6 +11,7 @@ const useProject = () => {
   const setSprintList = useSetRecoilState(sprintListState);
   const setUserList = useSetRecoilState(userListState);
   const setLabelList = useSetRecoilState(labelListState);
+  const setMindmap = useSetRecoilState(mindmapState);
 
   const roomId = useRoomId();
 
@@ -21,6 +24,15 @@ const useProject = () => {
     setSprintList(projectInfo.sprints);
     setUserList(projectInfo.users);
     setLabelList(projectInfo.labels);
+
+    const initNodes = new Map(
+      projectInfo.mindmap.map((node) => {
+        const { id: nodeId, ...props } = node;
+        return [nodeId, { nodeId, ...props }];
+      })
+    );
+    const rootId = projectInfo.mindmap.filter((node) => node.level === 'ROOT')[0].id;
+    setMindmap({ rootId: rootId as number, mindNodes: initNodes as IMindNodes });
   };
 
   useEffect(() => {
